Make the debian builds automatically generate the debian/changelog entries.
This makes it so we don't have to go update the changelog for each
release in every repository. We can't accomplish this the same way we do
for RPMs because we don't invoke the Makefile directly for debian
builds, which means we can't pass in the version info and it's not
available via environment variables. In order to get the full version
information right, this passes in both the version and the release number.