Do not set (and thus restrict) `-architectures`
The previous setting `-architectures=all` caused aptly to not upload
librdkafka, which has the architecture `amd64`.  To be on the safe side
for any further architecture-related issues, this commit removes this
  override (falling back to the default we have in aptly.conf.template,
which happens to be the default aptly value of `"architectures": []`).
The final effect is that ALL architectures are being included and thus
uploaded by aptly, including librdkafka.