Override confluent.maven.repo so we can build from our staging maven repository
This is requried for bootstrapping situations, e.g. when we are building
against an unreleased Kafka version (such as `0.9.0.0-cp1`);  here, our
various CP projects will need to know where to pull these artifacts from
and, unlike for `package.sh`, the `deploy_artifacts.sh` script does not
run within the packaging VMs and/or against a local maven repository
in which these artifacts might already be available.